+++ apps/c/airfoil/airfoil_plain_openclCPU/sp/res_calc_kernel.cpp
@@ -16,7 +16,7 @@ void op_x86_res_calc(
   float *ind_arg2,
   float *ind_arg3,
   int   *ind_map,
-  short *arg_map,
+  int *arg_map,
   int   *ind_arg_sizes,
   int   *ind_arg_offs,
   int    block_offset,
